Fixed CSS bugs... Improved SignalR interaction (Live changes to stickers).

+++ Kaberdin_PostItNotes/Data/Models/StickerModel.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Kaberdin_PostItNotes.Data.Models
 {
@@ -15,8 +12,8 @@ namespace Kaberdin_PostItNotes.Data.Models
 
         [Key]
         public int StickerId { get; set; }
-        public int? PositionX { get; set; }
-        public int? PositionY { get; set; }
+        public float? PositionX { get; set; }
+        public float? PositionY { get; set; }
         public int? Width { get; set; }
         public int? Height { get; set; }
         public StickerColor? Color { get; set; }
